refactor: add wire scripts
