+++ package.json
@@ -13,9 +13,10 @@
     "lint:fix": "eslint --fix '**/*.{js,ts,json}' && prettier --write . && solhint 'contracts/**/*.sol' --fix --noPrompt",
     "lint:js": "eslint '**/*.{js,ts,json}' && prettier --check .",
     "lint:sol": "solhint 'contracts/**/*.sol'",
+    "oapp:wire:mainnet": "npx hardhat lz:oapp:wire --oapp-config layerZero/mainnet/config.ts",
+    "oapp:wire:testnet": "npx hardhat lz:oapp:wire --oapp-config layerZero/testnet/config.ts",
     "prepare": "husky",
     "script:bridgeToken": "cross-env && forge script scripts/BridgeToken.s.sol --rpc-url ${NETWORK} --broadcast",
-    "script:setPeer": "cross-env && forge script scripts/SetPeer.s.sol --rpc-url ${NETWORK} --broadcast",
     "test": "forge test -vvv --nmt \"^.*invariant.*$\"",
     "test:invariant": "forge test --mt \"^.*invariant.*$\"",
     "verify": "npx hardhat verify --constructor-args utils/arguments.js"
